+++ src/main/java/kitchenpos/application/ProductService.java
@@ -7,7 +7,7 @@ import kitchenpos.dao.ProductRepository;
 import kitchenpos.domain.Product;
 import kitchenpos.dto.ProductRequest;
 import kitchenpos.dto.ProductResponse;
-import kitchenpos.vo.Money;
+import kitchenpos.support.money.Money;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
