+++ src/main/java/kitchenpos/domain/Menu.java
@@ -8,8 +8,8 @@ import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import kitchenpos.common.BaseEntity;
-import kitchenpos.vo.Money;
+import kitchenpos.support.domain.BaseEntity;
+import kitchenpos.support.money.Money;
 
 @Entity
 public class Menu extends BaseEntity {
