+++ src/main/java/kitchenpos/domain/MenuGroup.java
@@ -5,7 +5,7 @@ import static javax.persistence.GenerationType.IDENTITY;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import kitchenpos.common.BaseEntity;
+import kitchenpos.support.domain.BaseEntity;
 
 @Entity
 public class MenuGroup extends BaseEntity {
