+++ src/test/java/kitchenpos/common/MoneyConverterTest.java
@@ -3,7 +3,8 @@ package kitchenpos.common;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.math.BigDecimal;
-import kitchenpos.vo.Money;
+import kitchenpos.support.money.Money;
+import kitchenpos.support.money.MoneyConverter;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
