+++ src/test/java/kitchenpos/domain/MenuProductsTest.java
@@ -5,7 +5,7 @@ import static kitchenpos.fixture.ProductFixture.상품;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.List;
-import kitchenpos.vo.Money;
+import kitchenpos.support.money.Money;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
